+++ changelog.md
@@ -9,6 +9,7 @@ v1.4.14+1.4.14
 - Fix: auto scroll to the wrong position
 - Highlight selected article
 - Add loading animation to explorer
+- Add percent number increase animation
 
 v1.4.13+1.4.13